requirements.txt
# 爬虫逆向框架依赖
# ============ 核心依赖 ============
requests>=2.28.0
aiohttp>=3.8.0
curl_cffi>=0.5.0

# ============ 浏览器自动化 ============
playwright>=1.40.0
selenium>=4.8.0
webdriver-manager>=3.8.0

# ============ 数据处理 ============
beautifulsoup4>=4.12.0
lxml>=4.9.0
pyquery>=1.4.3
Pillow>=9.0.0
numpy>=1.24.0
orjson>=3.9.0

# ============ 数据库存储 ============
pymongo>=4.5.0
redis>=4.5.0
PyMySQL>=1.1.0

# ============ 加密/编码 ============
pycryptodome>=3.19.0
cryptography>=41.0.0

# ============ 异步支持 ============
asyncio>=3.4.3

# ============ WebSocket ============
websocket-client>=1.6.0

# ============ 日志/调试 ============
loguru>=0.7.0

# ============ 开发/测试 ============
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# ============ 代码质量 ============
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0

# ============ 类型扩展 ============
types-aiohttp>=3.8.0
types-redis>=4.5.0
